requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coinbase-commerce==1.0.1
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
stripe==7.2.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
